more sync gen code cleanup
